Move rev. swap methods in BoltzApi
