+++ libs/sdk-core/src/boltzswap.rs
@@ -3,10 +3,18 @@ use std::collections::HashMap;
 use serde::{Deserialize, Serialize};
 
 use anyhow::{anyhow, Result};
+use bitcoin_hashes::hex::ToHex;
+use bitcoin_hashes::{sha256, Hash};
+use serde_json::json;
 
 use const_format::concatcp;
+use rand::random;
+use reqwest::header::CONTENT_TYPE;
+use reqwest::Client;
 
 use crate::models::ReverseSwapInfo;
+use crate::reverseswap::CreateReverseSwapResponse;
+use crate::{ReverseSwap, ReverseSwapperAPI};
 
 const BOLTZ_API_URL: &str = "https://boltz.exchange/api/";
 const GET_PAIRS_ENDPOINT: &str = concatcp!(BOLTZ_API_URL, "getpairs");
@@ -77,6 +85,49 @@ struct Pairs {
     pairs: HashMap<String, Pair>,
 }
 
+pub struct BoltzApi {}
+
+#[tonic::async_trait]
+impl ReverseSwapperAPI for BoltzApi {
+    async fn reverse_swap_info(&self) -> Result<ReverseSwapInfo> {
+        reverse_swap_info().await
+    }
+
+    async fn create_reverse_swap(
+        &self,
+        amount_sat: u64,
+        onchain_claim_address: String,
+        pair_hash: String,
+        routing_node: String,
+    ) -> Result<ReverseSwap> {
+        let rand_bytes: [u8; 32] = random();
+        let preimage = sha256::Hash::hash(&rand_bytes);
+        let preimage_hash = sha256::Hash::hash(&preimage);
+        let preimage_hash_hex = preimage_hash.to_hex();
+
+        let response: CreateReverseSwapResponse = Client::new()
+            .post(CREATE_REVERSE_SWAP_ENDPOINT)
+            .header(CONTENT_TYPE, "application/json")
+            .body(get_boltz_reverse_swap_args(
+                amount_sat,
+                preimage_hash_hex,
+                pair_hash,
+                onchain_claim_address,
+                routing_node,
+            ))
+            .send()
+            .await?
+            .json()
+            .await?;
+
+        // TODO In case of error, return Err(str) or Ok(ReverseSwap{ error_message = ..} ) ?
+        return Ok(ReverseSwap {
+            error_message: None,
+            response,
+        });
+    }
+}
+
 pub async fn reverse_swap_info() -> Result<ReverseSwapInfo> {
     let pairs = reqwest::get(GET_PAIRS_ENDPOINT)
         .await?
@@ -97,4 +148,24 @@ pub async fn reverse_swap_info() -> Result<ReverseSwapInfo> {
             })
         }
     }
+}
+
+fn get_boltz_reverse_swap_args(
+    amount_sat: u64,
+    preimage_hash_hex: String,
+    pair_hash: String,
+    claim_pubkey: String,
+    routing_node: String,
+) -> String {
+    json!({
+        "type": "reversesubmarine",
+        "pairId": "BTC/BTC",
+        "orderSide": "buy",
+        "invoiceAmount": amount_sat,
+        "preimageHash": preimage_hash_hex,
+        "pairHash": pair_hash,
+        "claimPublicKey": claim_pubkey,
+        "routingNode": routing_node
+    })
+    .to_string()
 }
