+++ libs/sdk-core/src/reverseswap.rs
@@ -1,38 +1,9 @@
 use std::sync::Arc;
 
-use crate::boltzswap::CREATE_REVERSE_SWAP_ENDPOINT;
 use crate::chain::{ChainService, MempoolSpace};
-use anyhow::Result;
-use bitcoin_hashes::hex::ToHex;
-use bitcoin_hashes::{sha256, Hash};
-use rand::random;
-use reqwest::header::CONTENT_TYPE;
-use reqwest::Client;
-use serde::{Deserialize, Serialize};
-use serde_json::json;
-
-use crate::breez_services::BreezServer;
 use crate::models::{ReverseSwap, ReverseSwapperAPI};
-
-fn get_boltz_reverse_swap_args(
-    amount_sat: u64,
-    preimage_hash_hex: String,
-    pair_hash: String,
-    claim_pubkey: String,
-    routing_node: String,
-) -> String {
-    json!({
-        "type": "reversesubmarine",
-        "pairId": "BTC/BTC",
-        "orderSide": "buy",
-        "invoiceAmount": amount_sat,
-        "preimageHash": preimage_hash_hex,
-        "pairHash": pair_hash,
-        "claimPublicKey": claim_pubkey,
-        "routingNode": routing_node
-    })
-    .to_string()
-}
+use anyhow::Result;
+use serde::{Deserialize, Serialize};
 
 #[derive(Serialize, Deserialize, Debug)]
 #[serde(rename_all = "camelCase")]
@@ -45,49 +16,11 @@ pub struct CreateReverseSwapResponse {
     lockup_address: String,
 }
 
-#[tonic::async_trait]
-impl ReverseSwapperAPI for BreezServer {
-    async fn create_reverse_swap(
-        &self,
-        amount_sat: u64,
-        onchain_claim_address: String,
-        pair_hash: String,
-        routing_node: String,
-    ) -> Result<ReverseSwap> {
-        let rand_bytes: [u8; 32] = random();
-        let preimage = sha256::Hash::hash(&rand_bytes);
-        let preimage_hash = sha256::Hash::hash(&preimage);
-        let preimage_hash_hex = preimage_hash.to_hex();
-
-        let response_body = Client::new()
-            .post(CREATE_REVERSE_SWAP_ENDPOINT)
-            .header(CONTENT_TYPE, "application/json")
-            .body(get_boltz_reverse_swap_args(
-                amount_sat,
-                preimage_hash_hex,
-                pair_hash,
-                onchain_claim_address,
-                routing_node,
-            ))
-            .send()
-            .await?
-            .text()
-            .await?;
-        let response: CreateReverseSwapResponse = serde_json::from_str(&response_body)?;
-
-        // TODO In case of error, return Err(str) or Ok(ReverseSwap{ error_message = ..} ) ?
-        return Ok(ReverseSwap {
-            error_message: None,
-            response,
-        });
-    }
-}
-
 /// This struct is responsible for sending to an onchain address using lightning payments.
-/// It uses internally an implementation of ReverseSwapperAPI that represents Boltz reverse swapper service.
+/// It uses internally an implementation of [ReverseSwapperAPI] that represents Boltz reverse swapper service.
 pub(crate) struct BTCSendSwap {
     network: bitcoin::Network,
-    reverse_swapper_api: Arc<dyn ReverseSwapperAPI>,
+    pub(crate) reverse_swapper_api: Arc<dyn ReverseSwapperAPI>,
     persister: Arc<crate::persist::db::SqliteStorage>,
     chain_service: Arc<dyn ChainService>,
     //payment_sender: Arc<dyn Sender>,
